Start testing client
