+++ packages/client/src/ci-providers/__tests__/utils.ts
@@ -7,7 +7,3 @@ export function readEnvFile(path: string): Env {
   const envFile = readFileSync(path, "utf8");
   return dotenv.parse(envFile);
 }
-
-export function fixtureFactory<T>(defaults: T): (params?: Partial<T>) => T {
-  return (params = {}) => ({ ...defaults, ...params });
-}
